Bundle the filename into the io::location object.
I initially avoided this because of the overhead of storing the filename
in every single location object. However, having it separate makes
things harder to work with. The solution here is to allocate a name
which can be safely referred to by all location objects through a plain
pointer. This minimises the overhead of the location object while
solving the maintaibility problem.